requirements.txt
dash
plotly
fastapi
uvicorn
httpx
supabase
python-dotenv
openai
pandas
requests